+++ src/main/java/ru/oshifugo/functionalclans/listener/SalaryListener.java
@@ -16,7 +16,7 @@ import java.util.Map;
 
 public class SalaryListener implements Listener {
 
-    private final Map<String, Integer> clansTaxes = new HashMap<>();
+    private final Map<String, Long> clansTaxes = new HashMap<>();
 
     @EventHandler
     private void onSalaryEvent(OnSalaryEvent event) {
@@ -27,9 +27,9 @@ public class SalaryListener implements Listener {
         double clanTaxes = event.getAmount() / 100 * tax;
         event.setAmount(event.getAmount() - clanTaxes);
         event.setShouldSendMessage(false);
-        int clanCurrentCash = Clan.getCash(clanName);
-        Clan.setCash(clanName, String.valueOf(clanCurrentCash + (int) clanTaxes));
-        addTax(clanName, (int)clanTaxes);
+        long clanCurrentCash = Clan.getCash(clanName);
+        Clan.setCash(clanName, String.valueOf(clanCurrentCash + (long) clanTaxes));
+        addTax(clanName, (long)clanTaxes);
 //          "§cВы заработали §b%s₴. §cНалог клана: §b%d₴. §cИтого вы получаете: §a%d₴!",
         GUITranslatePlaceholder t = GUITranslate.getTranslate(event.getPlayer());
 
@@ -47,7 +47,7 @@ public class SalaryListener implements Listener {
             String clanName = Member.getClan(player.getName());
             if (clanName == null) continue;
             if (!clansTaxes.containsKey(clanName)) continue;
-            int tax = clansTaxes.get(clanName);
+            long tax = clansTaxes.get(clanName);
             if (tax == 0) continue;
             GUITranslatePlaceholder t = GUITranslate.getTranslate(player);
             player.sendMessage(t.get("salary-event.on-report")
@@ -57,8 +57,8 @@ public class SalaryListener implements Listener {
         clearTaxes();
     }
 
-    private void addTax(String clan, int tax) {
-        int clanTaxesNow = 0;
+    private void addTax(String clan, long tax) {
+        long clanTaxesNow = 0;
         if (clansTaxes.containsKey(clan)){
             clanTaxesNow = clansTaxes.get(clan);
         }
